refactor(metrics): Use consistent naming for devmetric mod
Rename block/block_metrics.rs to block/metrics.rs since it
makes more sense to deduce metrics from block (or other device)
namespaces.
