+++ src/vmm/src/devices/virtio/virtio_block/persist.rs
@@ -16,8 +16,8 @@ use super::*;
 use crate::devices::virtio::device::{DeviceState, IrqTrigger};
 use crate::devices::virtio::gen::virtio_blk::VIRTIO_BLK_F_RO;
 use crate::devices::virtio::persist::VirtioDeviceState;
-use crate::devices::virtio::virtio_block::block_metrics::BlockMetricsPerDevice;
 use crate::devices::virtio::virtio_block::device::FileEngineType;
+use crate::devices::virtio::virtio_block::metrics::BlockMetricsPerDevice;
 use crate::devices::virtio::TYPE_BLOCK;
 use crate::logger::warn;
 use crate::rate_limiter::persist::RateLimiterState;
