+++ src/vmm/src/devices/virtio/virtio_block/request.rs
@@ -15,8 +15,8 @@ pub use crate::devices::virtio::gen::virtio_blk::{
     VIRTIO_BLK_T_FLUSH, VIRTIO_BLK_T_GET_ID, VIRTIO_BLK_T_IN, VIRTIO_BLK_T_OUT,
 };
 use crate::devices::virtio::queue::DescriptorChain;
-use crate::devices::virtio::virtio_block::block_metrics::BlockDeviceMetrics;
 use crate::devices::virtio::virtio_block::device::DiskProperties;
+use crate::devices::virtio::virtio_block::metrics::BlockDeviceMetrics;
 use crate::logger::{error, IncMetric};
 use crate::rate_limiter::{RateLimiter, TokenType};
 use crate::vstate::memory::{ByteValued, Bytes, GuestAddress, GuestMemoryMmap};
