+++ src/vmm/src/logger/metrics.rs
@@ -73,7 +73,7 @@ use vm_superio::rtc_pl031::RtcEvents;
 
 use super::FcLineWriter;
 use crate::devices::virtio::net::metrics as net_metrics;
-use crate::devices::virtio::virtio_block::block_metrics;
+use crate::devices::virtio::virtio_block::metrics as block_metrics;
 #[cfg(target_arch = "aarch64")]
 use crate::warn;
